Migrate SetFieldValue method
